pyproject.toml
[tool.poetry]
name = "sing-along"
version = "0.1.0"
description = ""
authors = ["sam atkins <sam.d.atkins@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
psycopg2-binary = "^2.9.2"
channels = "^3.0.4"
waitress = "^2.1.2"
debugpy = "^1.6.0"
whitenoise = "^6.2.0"
black = "^22.10.0"
django-safedelete = "^1.3.1"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core=1.1.11"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "py"
url = "https://pypi.python.org/simple"